fixtwit fixes

+++ plugins/admin.py
@@ -73,11 +73,15 @@ async def deAmpMessage(triggerMessage):
 @commands.messageHandlerFilter("://twitter", filterType="cqc")
 async def fixTwit(triggerMessage):
     # only fix tweets if the embed failed to load
-    if not [embed.url for embed in triggerMessage.embeds if "/twitter" in embed.url]:
+    if not [embed.url for embed in triggerMessage.embeds if "/twitter" in embed.url and "/status/" in embed.url]:
         message = triggerMessage.content
         message = message.replace("://twitter", "://vxtwitter")
+
+        #advanced string manipulation time
+        #remove chaff
+        message = " ".join([x.split("?")[0] if "vxtwitter" in x.split(".com")[0] else x for x in message.split(" ")])
+        
         message = triggerMessage.author.name + ": " + message
-    
         await triggerMessage.channel.send(message)
         await triggerMessage.delete()
